# app/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Mock API Web Service",
    "description": "Mock API Web Service"
  },
  "host": "app.sidebox.es",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/echo/{status}": {
      "post": {
        "summary": "Echoes back request",
        "description": "Echoes headers, querystring, and request body",
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "description": "Status code to be returned",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "echoRequest",
            "description": "Request body to be echoed",
            "schema": {
              "$ref": "#/definitions/echoRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/echoResponse"
            }
          }
        }
      }
    },
    "/echo-from-text/{status}": {
      "post": {
        "summary": "Echoes text as json",
        "description": "Echoes headers, querystring, and request body",
        "consumes": [
          "text/plain"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "description": "Status code to be returned",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "echoRequest",
            "x-ntx-summary": "Text request body to be echoed as json",
            "description": "{\"abc\": 123, \"def\": true, \"ghi\": \"value1\"}",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/echoResponse"
            }
          }
        }
      }
    },
    "/files/errors/{status}": {
      "get": {
        "summary": "Download error",
        "description": "Returns error code instead of image for download",
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "description": "Error status code to be returned",
            "required": true,
            "type": "integer"
          }
        ],
        "produces": [
          "images/jpeg"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "file": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          }
        }
      }
    },
...
    "/files/upload/base64": {
      "post": {
        "summary": "Upload file as base64",
        "description": "Upload file as base64",
        "parameters": [
          {
            "in": "body",
            "name": "rootBody",
            "description": "Root Body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "fileContent": {
                  "description": "File to upload",
                  "x-ntx-summary": "File to upload",
                  "type": "string",
                  "format": "byte"
                },
                "customName": {
                  "description": "Custom name with extension",
                  "x-ntx-summary": "Custom name with extension",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "customName": {
                  "type": "string",
                  "description": "Custom file name Output",
                  "x-ntx-summary": "Custom file name Output"
                },
                "mimeType": {
                  "type": "string",
                  "description": "MIME type Output",
                  "x-ntx-summary": "MIME type Output"
                },
                "md5": {
                  "type": "string",
                  "description": "MD5 Output",
                  "x-ntx-summary": "MD5 Output"
                },
                "size": {
                  "type": "integer",
                  "description": "File size Output",
                  "x-ntx-summary": "File size Output"
                }
              }
            }
          }
        }
      }
    },
    "/files/upload/base64/multi": {
      "post": {
        "summary": "Upload multiple files as base64",
        "description": "Upload multiple files as base64",
        "parameters": [
          {
            "in": "body",
            "name": "rootBody",
            "description": "Root Body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "fileContent": {
                    "description": "File to upload",
                    "x-ntx-summary": "File to upload",
                    "type": "string",
                    "format": "byte"
                  },
                  "customName": {
                    "description": "Custom name with extension",
                    "x-ntx-summary": "Custom name with extension",
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer",
                  "description": "Number of files",
                  "x-ntx-summary": "Number of files"
                },
                "files": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "customName": {
                        "type": "string",
                        "description": "Custom file name Output",
                        "x-ntx-summary": "Custom file name Output"
                      },
                      "mimeType": {
                        "type": "string",
                        "description": "MIME type Output",
                        "x-ntx-summary": "MIME type Output"
                      },
                      "md5": {
                        "type": "string",
                        "description": "MD5 Output",
                        "x-ntx-summary": "MD5 Output"
                      },
                      "size": {
                        "type": "integer",
                        "description": "File size Output",
                        "x-ntx-summary": "File size Output"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/files/upload/form-data": {
      "post": {
        "summary": "Upload file as multipart/form-data",
        "description": "Upload file as multipart/form-data",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file1",
            "description": "File to upload",
            "required": true,
            "type": "file"
          },
          {
            "in": "formData",
            "name": "customName",
            "description": "Custom name with extension",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "originalName": {
                  "type": "string",
                  "description": "Original file name Output",
                  "x-ntx-summary": "Original file name Output"
                },
                "customName": {
                  "type": "string",
                  "description": "Custom file name Output",
                  "x-ntx-summary": "Custom file name Output"
                },
                "mimeType": {
                  "type": "string",
                  "description": "MIME type Output",
                  "x-ntx-summary": "MIME type Output"
                },
                "md5": {
                  "type": "string",
                  "description": "MD5 Output",
                  "x-ntx-summary": "MD5 Output"
                },
                "size": {
                  "type": "integer",
                  "description": "File size Output",
                  "x-ntx-summary": "File size Output"
                }
              }
            }
          }
        }
      }
    },
...
    "/files/upload/octet-stream": {
      "post": {
        "summary": "Upload file as octet-stream",
        "description": "Upload file as octet-stream",
        "consumes": [
          "application/octet-stream"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "file1",
            "description": "File to upload",
            "required": true,
            "schema": {
              "type": "string",
              "format": "binary"
            }
          },
          {
            "in": "header",
            "name": "Custom-Name",
            "x-ntx-summary": "Custom name with extension",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "customName": {
                  "type": "string",
                  "description": "Custom file name Output",
                  "x-ntx-summary": "Custom file name Output"
                },
                "mimeType": {
                  "type": "string",
                  "description": "MIME type Output",
                  "x-ntx-summary": "MIME type Output"
                },
                "md5": {
                  "type": "string",
                  "description": "MD5 Output",
                  "x-ntx-summary": "MD5 Output"
                },
                "size": {
                  "type": "integer",
                  "description": "File size Output",
                  "x-ntx-summary": "File size Output"
                }
              }
            }
          }
        }
      }
    },
...
    "/files/upload/uri": {
      "post": {
        "summary": "Upload file as URI",
        "description": "Upload file as URI",
        "parameters": [
          {
            "in": "body",
            "name": "rootBody",
            "description": "Root Body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "fileUri": {
                  "description": "File to upload",
                  "x-ntx-summary": "File to upload",
                  "type": "string",
                  "format": "x-ntx-file-reference"
                },
                "customName": {
                  "description": "Custom name with extension",
                  "x-ntx-summary": "Custom name with extension",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "customName": {
                  "type": "string",
                  "description": "Custom file name Output",
                  "x-ntx-summary": "Custom file name Output"
                },
                "mimeType": {
                  "type": "string",
                  "description": "MIME type Output",
                  "x-ntx-summary": "MIME type Output"
                },
                "md5": {
                  "type": "string",
                  "description": "MD5 Output",
                  "x-ntx-summary": "MD5 Output"
                },
                "size": {
                  "type": "integer",
                  "description": "File size Output",
                  "x-ntx-summary": "File size Output"
                }
              }
            }
          }
        }
      }
    },
    "/files/upload/uri/multi": {
      "post": {
        "summary": "Upload multiple files as URI",
        "description": "Upload multiple files as URI",
        "parameters": [
          {
            "in": "body",
            "name": "rootBody",
            "description": "Root Body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "fileUri": {
                    "description": "File to upload",
                    "x-ntx-summary": "File to upload",
                    "type": "string",
                    "format": "x-ntx-file-reference"
                  },
                  "customName": {
                    "description": "Custom name with extension",
                    "x-ntx-summary": "Custom name with extension",
                    "type": "string"
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "count": {
                  "type": "integer",
                  "description": "Number of files",
                  "x-ntx-summary": "Number of files"
                },
                "files": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "customName": {
                        "type": "string",
                        "description": "Custom file name Output",
                        "x-ntx-summary": "Custom file name Output"
                      },
                      "mimeType": {
                        "type": "string",
                        "description": "MIME type Output",
                        "x-ntx-summary": "MIME type Output"
                      },
                      "md5": {
                        "type": "string",
                        "description": "MD5 Output",
                        "x-ntx-summary": "MD5 Output"
                      },
                      "size": {
                        "type": "integer",
                        "description": "File size Output",
                        "x-ntx-summary": "File size Output"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/all-types": {
      "post": {
        "summary": "POST all variable types",
        "description": "POST all variable types",
        "parameters": [
          {
            "in": "body",
            "name": "rootBody",
            "required": true,
            "schema": {
              "$ref": "#/definitions/rootBody"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/allTypesOutputs"
            }
          }
        }
      }
    },
    "/all-types/object": {
      "get": {
        "summary": "Returns hardcoded object",
        "description": "Returns hardcoded object and text",
        "parameters": [
          {
            "in": "query",
            "name": "expected",
            "description": "Expected data to be returned",
            "required": false,
            "type": "string",
            "default": "body",
            "enum": [
              "body",
              "empty",
              "plaintext"
            ]
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/allTypesOutputsObject"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "echoRequest": {
      "type": "object"
    },
    "echoResponse": {
      "type": "object"
    },
    "rootBody": {
      "type": "object",
      "properties": {
        "allTypesInputs": {
          "type": "object",
          "properties": {
            "textInput": {
              "type": "string",
              "description": "Text Input",
              "x-ntx-summary": "Text Input"
            },
            "decimalInput": {
              "type": "number",
              "description": "Decimal Input",
              "x-ntx-summary": "Decimal Input"
            },
            "integerInput": {
              "type": "integer",
              "description": "Integer Input",
              "x-ntx-summary": "Integer Input"
            },
            "booleanInput": {
              "type": "boolean",
              "description": "Boolean Input",
              "x-ntx-summary": "Boolean Input"
            },
            "datetimeInput": {
              "type": "string",
              "format": "date-time",
              "description": "Datetime Input",
              "x-ntx-summary": "Datetime Input"
            },
            "collectionInput": {
              "type": "array",
              "items": {
                "type": "string",
                "description": "Collection Item Input"
              },
              "description": "Collection Input",
              "x-ntx-summary": "Collection Input"
            }
          }
        }
      }
    },
...
  }
}